Addition of ordering field for member positions. Changes in manifest and installation routine. Addition of ACL rules for research areas and members

+++ src/admin/models/publications/publication.php
@@ -35,34 +35,35 @@ class JResearchAdminModelPublication extends JModelForm{
         {
             if (empty($this->data))
             {
-                    $app = & JFactory::getApplication();
-                    $data = & JRequest::getVar('jform');
-                    if (empty($data))
-                    {
-                            $selected = & JRequest::getVar('cid', 0, '', 'array');
-                            $db = JFactory::getDBO();
-                            $query = $db->getQuery(true);
-                            // Select all fields from the hello table.
-                            $query->select('*');
-                            $query->from('`#__jresearch_publication`');
-                            $query->where('id = ' . (int)$selected[0]);
-                            $db->setQuery((string)$query);
-                            $data = & $db->loadAssoc();
-                    }
-                    if (empty($data))
-                    {
-						// Check the session for previously entered form data.
-                        $data = $app->getUserState('com_jresearch.edit.publication.data', array());
-                    }
+            	$app = & JFactory::getApplication();
+                $data = & JRequest::getVar('jform');
+                if (empty($data))
+                {
+                	$selected = & JRequest::getVar('cid', 0, '', 'array');
+                    $db = JFactory::getDBO();
+                    $query = $db->getQuery(true);
+                    // Select all fields from the hello table.
+                    $query->select('*');
+                    $query->from('`#__jresearch_publication`');
+                    $query->where('id = ' . (int)$selected[0]);
+                    $db->setQuery((string)$query);
+                    $data = & $db->loadAssoc();
+                }
+                if (empty($data))
+                {
+					// Check the session for previously entered form data.
+                	$data = $app->getUserState('com_jresearch.edit.publication.data', array());
+                }
                     
-                    //Once the data is retrieved, time to fix it
-                    if(is_string($data['id_research_area'])){
-                    	$data['id_research_area'] = explode(',', $data['id_research_area']);
-                    }
+                //Once the data is retrieved, time to fix it
+                if(is_string($data['id_research_area'])){
+                	$data['id_research_area'] = explode(',', $data['id_research_area']);
+                }
                     
-                    $app->setUserState('com_jresearch.edit.publication.data', $data);
-                    $this->data = $data;
+                $app->setUserState('com_jresearch.edit.publication.data', $data);
+                $this->data = $data;
             }
+            
             return $this->data;
         }
 
@@ -88,97 +89,120 @@ class JResearchAdminModelPublication extends JModelForm{
          */
         function save()
         {
-				$app = JFactory::getApplication();
-				jresearchimport('helpers.publications', 'jresearch.admin');
-				$params = JComponentHelper::getParams('com_jresearch');
-				$omittedFields = array();
+			$app = JFactory::getApplication();
+			jresearchimport('helpers.publications', 'jresearch.admin');
+			$params = JComponentHelper::getParams('com_jresearch');
+			$omittedFields = array();
                 
-                $data =& $this->getData();                
-                $row =& $this->getTable('Publication', 'JResearch');
+            $data =& $this->getData();                
+            $row =& $this->getTable('Publication', 'JResearch');
 
-                //Time to upload the file
-                $delete = $data['delete_url'];
-	    		if($delete == 1){
-	    			if(!empty($data['files'])){
-		    			$filetoremove = JRESEARCH_COMPONENT_ADMIN.DS.$params->get('files_root_path', 'files').DS.'publications'.DS.$row->files;
-		    			$data['files'] = '';
-		    			@unlink($filetoremove);
-	    			}
-		    	}
-		    		    
-	    		$files = JRequest::getVar('jform', array(), 'FILES');
-			    if(!empty($files['name']['file_url'])){	    	
-		    		$data['files'] = JResearchUtilities::uploadDocument($files, 'file_url', $params->get('files_root_path', 'files').DS.'publications');
+            //Time to upload the file
+            $delete = $data['delete_url'];
+	    	if($delete == 1){
+	    		if(!empty($data['files'])){
+		    		$filetoremove = JRESEARCH_COMPONENT_ADMIN.DS.$params->get('files_root_path', 'files').DS.'publications'.DS.$row->files;
+		    		$data['files'] = '';
+		    		@unlink($filetoremove);
 	    		}
+		    }
+		    		    
+	    	$files = JRequest::getVar('jform', array(), 'FILES');
+			if(!empty($files['name']['file_url'])){	    	
+		    	$data['files'] = JResearchUtilities::uploadDocument($files, 'file_url', $params->get('files_root_path', 'files').DS.'publications');
+	    	}
 	    		
-	    		if($data['resethits'] == 1){
-			    	$data['hits'] = 0;
-			    }else{
-			    	$omittedFields[] = 'hits';			    	
-			    }
+	    	if($data['resethits'] == 1){
+				$data['hits'] = 0;
+			}else{
+				$omittedFields[] = 'hits';			    	
+			}
 			    			    			    
-			    //Now time for the authors
-				$maxAuthors = (int)$data['nauthorsfield'];
-				$authorsArray = array();
-				for($j = 0; $j < $maxAuthors; $j++){
-					$value = $data["authorsfield".$j];
-					if(!empty($value)){
-						$row->addAuthor($value);
-					}
+			//Now time for the authors
+			$maxAuthors = (int)$data['nauthorsfield'];
+			$authorsArray = array();
+			for($j = 0; $j < $maxAuthors; $j++){
+				$value = $data["authorsfield".$j];
+				if(!empty($value)){
+					$row->addAuthor($value);
 				}
+			}
 				
-				$data['authors'] = $row->authors;
+			$data['authors'] = $row->authors;
 				
-				//Citekey generation
-				$data['citekey'] = JResearchPublicationsHelper::generateCitekey($data);
+			//Citekey generation
+			$data['citekey'] = JResearchPublicationsHelper::generateCitekey($data);
 				
-				//Alias generation
-				if(empty($data['alias'])){
-					$data['alias'] = JFilterOutput::stringURLSafe($data['title']);
-				}
+			//Alias generation
+			if(empty($data['alias'])){
+				$data['alias'] = JFilterOutput::stringURLSafe($data['title']);
+			}
 				
-				//Checking of research areas
-				if(!empty($data['id_research_area'])){
-					if(in_array('1', $data['id_research_area'])){
-						$data['id_research_area'] = '1';
-					}else{
-						$data['id_research_area'] = implode(',', $data['id_research_area']);
-					}
-				}else{
+			//Checking of research areas
+			if(!empty($data['id_research_area'])){
+				if(in_array('1', $data['id_research_area'])){
 					$data['id_research_area'] = '1';
+				}else{
+					$data['id_research_area'] = implode(',', $data['id_research_area']);
 				}
+			}else{
+				$data['id_research_area'] = '1';
+			}
 				
 				
-                if (!$row->save($data, '', $omittedFields))
-                {
-                	//Since the save routine modifies the array data
-        			JRequest::setVar('jform', $data);                	
-                    $this->setError($row->getError());
-                    return false;
-                }
+            if (!$row->save($data, '', $omittedFields))
+            {
+            	//Since the save routine modifies the array data
+        		JRequest::setVar('jform', $data);                	
+                $this->setError($row->getError());
+                return false;
+            }
 
-                $data['id'] = $row->id;
-                $app->setUserState('com_jresearch.edit.publication.data', $data);
+            $data['id'] = $row->id;
+            $app->setUserState('com_jresearch.edit.publication.data', $data);
 
-                return true;
+            return true;
         }
         
         /**
          * Publishes the set of selected items
          */
         function publish(){
-           $selected =& JRequest::getVar('cid', 0, '', 'array');
-           $publication = JTable::getInstance('Publication', 'JResearch');
-           return $publication->publish($selected, 1);
+           $selected = & JRequest::getVar('cid', 0, '', 'array');
+	       $publication = JTable::getInstance('Publication', 'JResearch');           
+	       $allOk = true;
+           foreach($selected as $id){
+           	   $action = JResearchAccessHelper::getActions('publication', $id);           	
+           	   if($action->get('core.publications.edit.state')){
+	    	       $allOk = $allOk && $publication->publish($id, 1);
+           	   }else{
+           	   	   $allOk = false;
+           		   $this->setError(JText::sprintf('JRESEARCH_EDIT_ITEM_STATE_NOT_ALLOWED', $id));           	   	   
+           	   }
+           }
+           
+           return $allOk;
+        	
         }
 
         /**
          * Unpublishes the set of selected items
          */
         function unpublish(){
-           $selected =& JRequest::getVar('cid', 0, '', 'array');
-           $publication = JTable::getInstance('Publication', 'JResearch');
-           return $publication->publish($selected, 0);
+		   $selected = & JRequest::getVar('cid', 0, '', 'array');
+	       $publication = JTable::getInstance('Publication', 'JResearch');           
+	       $allOk = true;
+           foreach($selected as $id){
+           	   $action = JResearchAccessHelper::getActions('publication', $id);           	
+           	   if($action->get('core.publications.edit.state')){
+	    	       $allOk = $allOk && $publication->publish($id, 0);
+           	   }else{
+           	   	   $allOk = false;
+           		   $this->setError(JText::sprintf('JRESEARCH_EDIT_ITEM_STATE_NOT_ALLOWED', $id));           	   	   
+           	   }
+           }
+           
+           return $allOk;
 
         }
 
@@ -191,14 +215,19 @@ class JResearchAdminModelPublication extends JModelForm{
            $n = 0;
            $selected =& JRequest::getVar('cid', 0, '', 'array');
            $publication = JTable::getInstance('Publication', 'JResearch');
-           $user = JFactory::getUser();
+           $user = JFactory::getUser();           
            foreach($selected as $id){
-                $publication->load($id);
-	           	if(!$publication->isCheckedOut($user->get('id'))){	
-                	if($publication->delete($id)){
-            	        $n++;
-                	}
-	           	}
+        		$actions = JResearchAccessHelper::getActions('publication', $id);
+        		if($actions->get('core.publications.delete')){           	
+	                $publication->load($id);
+		           	if(!$publication->isCheckedOut($user->get('id'))){	
+        	        	if($publication->delete($id)){
+            		        $n++;
+                		}
+	           		}
+        		}else{
+        			$this->setError(JText::sprintf('JRESEARCH_DELETE_ITEM_NOT_ALLOWED', $id));
+        		}
            }
                                  
            return $n;           
@@ -243,8 +272,13 @@ class JResearchAdminModelPublication extends JModelForm{
         	$user = JFactory::getUser();
         	$n = 0;
         	foreach($cid as $id){
-        		if($publication->toggleInternal(array($id), $value, $user->get('id'))){
-        			$n++;
+        		$actions = JResearchAccessHelper::getActions('publication', $id);
+        		if($actions->get('core.publications.edit.state')){
+	        		if($publication->toggleInternal(array($id), $value, $user->get('id'))){
+    	    			$n++;
+        			}
+        		}else{
+        			$this->setError(JText::sprintf('JRESEARCH_EDIT_ITEM_STATE_NOT_ALLOWED', $id));
         		}
         	}
         	
@@ -259,19 +293,22 @@ class JResearchAdminModelPublication extends JModelForm{
         public function toggleInternal(){
         	$publication = $this->getTable('Publication', 'JResearch');
         	$cid = JRequest::getVar('cid', array());
-        	$publication->load($cid[0]);
-        	if(!empty($publication->id)){
-        		$user = JFactory::getUser();
-        		if(!$publication->isCheckedOut($user->get('id'))){
-	        		$publication->internal = !$publication->internal;	
-	        		return $publication->store();
-        		}else{
-        			return false;
-        		}
-        	}
-        	
-        	return false;
-        }
+			if(!empty($cid)){
+	        	$actions = JResearchAccessHelper::getActions('publication', $cid[0]);
+    	    	if($actions->get('core.publications.edit.state')){
+		        	$publication->load($cid[0]);
+		        	if(!empty($publication->id)){
+	    	    		$user = JFactory::getUser();
+	        			if(!$publication->isCheckedOut($user->get('id'))){
+		        			$publication->internal = !$publication->internal;	
+		        			return $publication->store();
+	        			}
+	        		}
+    	    	}
+			}
+			
+			return false;
+	    }
 	
 }
 ?>
